[ETCM-271] Minor fixes including packages renaming and FIXMEs for Mantis txs support
