+++ conseil-api/src/main/resources/metadata/mantis.testnet.conf
@@ -1,4 +1,3 @@
-// Important: Changing this metadata file, will affect Quorum Blockchain too
 {
   entities {
     blocks {
