--- conseil-common/src/main/scala/tech/cryptonomic/conseil/common/EvmDomain/Opcode.scala
@@ -1,6 +1,0 @@
-package tech.cryptonomic.conseil.common.EvmDomain
-
-/**
-  * Ethereum bytecode's opcode.
-  */
-case class Opcode(offset: Int, instruction: Instruction, parameters: BigInt)
