+++ conseil-common/src/main/scala/tech/cryptonomic/conseil/common/evm/domain/Instruction.scala
@@ -1,4 +1,4 @@
-package tech.cryptonomic.conseil.common.EvmDomain
+package tech.cryptonomic.conseil.common.evm.domain
 
 /**
   * Contract opcode instruction.
