+++ conseil-common/src/main/scala/tech/cryptonomic/conseil/common/evm/domain/Opcode.scala
@@ -1,0 +1,3 @@
+package tech.cryptonomic.conseil.common.evm.domain
+
+case class Opcode(offset: Int, instruction: Instruction, parameters: BigInt)
