+++ conseil-common/src/main/scala/tech/cryptonomic/conseil/common/evm/domain/Token.scala
@@ -1,8 +1,5 @@
-package tech.cryptonomic.conseil.common.EvmDomain
+package tech.cryptonomic.conseil.common.evm.domain
 
-/**
-  * Ethereum token.
-  */
 case class Token(
     address: String,
     blockHash: String,
