+++ conseil-common/src/main/scala/tech/cryptonomic/conseil/common/mantis/MantisPersistence.scala
@@ -7,23 +7,19 @@ import cats.Id
 import cats.effect.{Concurrent, Resource}
 import com.typesafe.scalalogging.LazyLogging
 import slick.jdbc.PostgresProfile.api._
-
+import tech.cryptonomic.conseil.common.evm.domain.{Contract, Token}
 import tech.cryptonomic.conseil.common.util.Conversion
 import tech.cryptonomic.conseil.common.util.Conversion.Syntax._
 import tech.cryptonomic.conseil.common.util.HexUtil.hexStringToBigDecimal
 import tech.cryptonomic.conseil.common.mantis.rpc.json.{Block, Log, Transaction}
 import tech.cryptonomic.conseil.common.mantis.MantisPersistence._
 import tech.cryptonomic.conseil.common.mantis.rpc.json.TransactionReceipt
-import tech.cryptonomic.conseil.common.EvmDomain.{Contract, Token}
 
-
-/**
-  * Mantis persistence into the database using Slick.
-  */
 class MantisPersistence[F[_]: Concurrent] extends LazyLogging {
 
   /**
     * SHA-3 signature for: Transfer(address,address,uint256)
+    * FIXME: verify and fix compatibility with Mantis - current implementation is for Ethereum
     */
   private val tokenTransferSignature = "0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef"
 
@@ -57,6 +53,8 @@ class MantisPersistence[F[_]: Concurrent] extends LazyLogging {
     * Create [[DBIO]] seq with contracts.
     *
     * @param logs JSON_RPC contract
+    *
+    * FIXME: verify and fix compatibility with Mantis - current implementation is for Ethereum
     */
   def createContracts(contracts: List[Contract]) =
     DBIO.seq(
@@ -67,6 +65,8 @@ class MantisPersistence[F[_]: Concurrent] extends LazyLogging {
     * Create [[DBIO]] seq with tokens.
     *
     * @param logs JSON_RPC token
+    *
+    * FIXME: verify and fix compatibility with Mantis - current implementation is for Ethereum
     */
   def createTokens(tokens: List[Token]) =
     DBIO.seq(
@@ -140,6 +140,7 @@ object MantisPersistence {
     * Convert form [[Transaction]] to [[Tables.TransactionsRow]]
     * TODO: This conversion should be done with the Chimney,
     *       but it's blocked due to the https://github.com/scala/bug/issues/11157
+    * FIXME: change to Mantis transaction (current implementation is for Ethereum)
     */
   implicit val transactionToTransactionsRow: Conversion[Id, Transaction, Tables.TransactionsRow] =
     new Conversion[Id, Transaction, Tables.TransactionsRow] {
@@ -166,6 +167,7 @@ object MantisPersistence {
     * Convert form [[TransactionReceipt]] to [[Tables.ReceiptsRow]]
     * TODO: This conversion should be done with the Chimney,
     *       but it's blocked due to the https://github.com/scala/bug/issues/11157
+    * FIXME: change to Mantis transaction receipt (current implementation is for Ethereum)
     */
   implicit val transactionReceiptToReceiptsRow: Conversion[Id, TransactionReceipt, Tables.ReceiptsRow] =
     new Conversion[Id, TransactionReceipt, Tables.ReceiptsRow] {
@@ -188,6 +190,7 @@ object MantisPersistence {
     * Convert form [[Log]] to [[Tables.LogsRow]]
     * TODO: This conversion should be done with the Chimney,
     *       but it's blocked due to the https://github.com/scala/bug/issues/11157
+    * FIXME: adapt to Mantis logs (current implementation is for Ethereum)
     */
   implicit val logToLogsRow: Conversion[Id, Log, Tables.LogsRow] =
     new Conversion[Id, Log, Tables.LogsRow] {
@@ -209,6 +212,7 @@ object MantisPersistence {
     * Convert form [[Contract]] to [[Tables.ContractsRow]]
     * TODO: This conversion should be done with the Chimney,
     *       but it's blocked due to the https://github.com/scala/bug/issues/11157
+    * FIXME: adapt to Mantis contracts (current implementation is for Ethereum)
     */
   implicit val contractToContractsRow: Conversion[Id, Contract, Tables.ContractsRow] =
     new Conversion[Id, Contract, Tables.ContractsRow] {
@@ -227,6 +231,7 @@ object MantisPersistence {
     * Convert form [[Log]] to [[Tables.TokenTransfersRow]]
     * TODO: This conversion should be done with the Chimney,
     *       but it's blocked due to the https://github.com/scala/bug/issues/11157
+    * FIXME: adapt to Mantis Tokens (current implementation is for Ethereum)
     */
   implicit val logToTokenTransfersRow: Conversion[Id, Log, Tables.TokenTransfersRow] =
     new Conversion[Id, Log, Tables.TokenTransfersRow] {
@@ -244,6 +249,7 @@ object MantisPersistence {
     * Convert form [[Token]] to [[Tables.TokensRow]]
     * TODO: This conversion should be done with the Chimney,
     *       but it's blocked due to the https://github.com/scala/bug/issues/11157
+    * FIXME: adapt to Mantis Tokens (current one is for Ethereum)
     */
   implicit val tokenToTokensRow: Conversion[Id, Token, Tables.TokensRow] =
     new Conversion[Id, Token, Tables.TokensRow] {
