+++ conseil-common/src/main/scala/tech/cryptonomic/conseil/common/mantis/MantisTypes.scala
@@ -2,7 +2,6 @@ package tech.cryptonomic.conseil.common.mantis
 
 object MantisTypes {
 
-  /** Case class representing hash to identify blocks across many block chains */
   final case class MantisBlockHash(value: String) extends AnyVal
 
 }
