+++ conseil-lorre/src/main/scala/tech/cryptonomic/conseil/indexer/mantis/MantisOperations.scala
@@ -4,13 +4,12 @@ import cats.effect.{Concurrent, Resource}
 import com.typesafe.scalalogging.LazyLogging
 import fs2.Stream
 import slickeffect.Transactor
-
 import tech.cryptonomic.conseil.common.config.Platforms.MantisBatchFetchConfiguration
+import tech.cryptonomic.conseil.common.evm.domain.{Bytecode, Contract}
 import tech.cryptonomic.conseil.common.rpc.RpcClient
 import tech.cryptonomic.conseil.common.mantis.MantisPersistence
 import tech.cryptonomic.conseil.common.mantis.rpc.MantisClient
 import tech.cryptonomic.conseil.indexer.config.{Custom, Depth, Everything, Newest}
-import tech.cryptonomic.conseil.common.EvmDomain.{Bytecode, Contract}
 
 /**
   * Mantis operations for Lorre.
